Add files via upload
A rotation angle is optional and can be entered as the last command line argument. The dragon curve is exactly the same as before if no rotation angle is entered in the command line. Although if a rotation angle is entered the dragon curve will be modified. The dragon curve looks coolest if the rotation angle is "1.5" or very close to the standard rotation. I would recommend the rotation angle to be within approximately 1.0 of the standard of "1.5707" in order for it to still resemble the dragon curve, but we could allow for larger or smaller rotation angles for experimentation if you guys want?

I wrote this on line 111

if (rotation_angle < 1.45) or (rotation_angle > 1.63):
		print >> sys.stderr, 'To resemble the standard dragon curve the rotation angle should be  1.45 < r < 1.63'
		print >> sys.stderr, 'Rotation angle r can go out of the recommended range for experimentation'
		print >> sys.stderr, 'Inorder to make the imager clearer as the rotation angle increases the line len